First-pass strawman of a kubectl shell
We should get Ben's bash-driven stuff in here, too. It may prove the
best option.

Still need to get the 'cd' command working and figure out dynamic
completion (that which can't be derived from cobra.Command api)

Oh, and in case it's not obvious when you try to run this, it still
needs some work.