Stop panics from leaking out of kubectl and crashing our shell (fixes #8)
Our cmdutil.BehaviorOnFatal was flawed in that it only printed errors
but did not panic to interrupt control flow of the calling
functions. This meant that normally fatal errors were not causing the
program to terminate with a nice error message and instead proceed to
create nasty stack traces.

Since we don't actually want kubesh to terminate, we now panic in
BehaviorOnFatal and recover from that panic gracefully. This allows
the expected error messages to get printed without any stack traces or
abrupt termination of kubesh.