Implement a tokenizer for better input handling
We were splitting on single spaces, which worked poorly with multiple
spaces, tabs, escaped spaces, or spaces within quotes.

This implementation parses the input by walking each rune, keeping state
as it goes.

This doesn't yet handle newlines, and there may be a better
implementation waiting to be built on top of Go's text/scanner.

Related to #18, possibly as a fix, pending the opinion of others.