Refactored to use cobra for our internal commands
The main fallout from this is the introduction of the --clear option to
reset the pin. I don't love this, but without it, we either need 1) a
special case for the pin command when applying the context or b) we need
to account for the context when setting it. I loved both of those
slightly less than --clear, but I'm not married to it.